+++ src/main/java/gift/controller/ProductController.java
@@ -3,8 +3,8 @@ package gift.controller;
 import gift.dto.ProductRequest;
 import gift.dto.ProductResponse;
 import gift.model.MemberRole;
-import gift.service.PageService;
 import gift.service.ProductService;
+import gift.service.page.PageService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
