+++ src/main/java/gift/controller/WishProductController.java
@@ -3,8 +3,8 @@ package gift.controller;
 import gift.dto.WishProductAddRequest;
 import gift.dto.WishProductResponse;
 import gift.dto.WishProductUpdateRequest;
-import gift.service.PageService;
 import gift.service.WishProductService;
+import gift.service.page.PageService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
